+++ modules/concurrency/include/hephaestus/concurrency/io_ring/io_ring.h
@@ -13,7 +13,7 @@
 #include <liburing/io_uring.h>
 #include <stdexec/stop_token.hpp>
 
-#include "hephaestus/concurrency/io_ring_operation_pointer.h"
+#include "hephaestus/concurrency/io_ring/io_ring_operation_pointer.h"
 
 namespace heph::concurrency {
