+++ modules/concurrency/include/hephaestus/concurrency/io_ring/stoppable_io_ring_operation.h
@@ -12,8 +12,8 @@
 #include <liburing/io_uring.h>
 #include <stdexec/stop_token.hpp>
 
-#include "hephaestus/concurrency/io_ring.h"
-#include "hephaestus/concurrency/io_ring_operation_pointer.h"
+#include "hephaestus/concurrency/io_ring/io_ring.h"
+#include "hephaestus/concurrency/io_ring/io_ring_operation_pointer.h"
 #include "hephaestus/utils/exception.h"
 
 namespace heph::concurrency {
